releases/requirements.txt
# Core Dependencies
anthropic>=0.18.0
python-binance>=1.0.19
ccxt>=4.2.0
requests>=2.31.0
websocket-client>=1.7.0
packaging>=23.0

# GUI
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
pyqtgraph>=0.13.3
matplotlib>=3.8.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
TA-Lib>=0.4.28

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.15.0  # Optional, for advanced ML
torch>=2.1.0  # Optional, alternative to TensorFlow

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
cryptography>=41.0.0
pyyaml>=6.0.0
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging
loguru>=0.7.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# Build
pyinstaller>=6.3.0
